+++ src/app/components/baas-components/network/network.module.ts
@@ -7,22 +7,22 @@ import { BuilderComponent } from "./builder/builder.component";
 import { MatPaginatorModule } from "@angular/material/paginator";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { MatTableModule } from "@angular/material/table";
-import { ReactiveFormsModule } from "@angular/forms";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatInputModule } from "@angular/material/input";
 import { MatCheckboxModule } from "@angular/material/checkbox";
 import { MatTabsModule } from "@angular/material/tabs";
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 
 @NgModule({
   declarations: [ManagementComponent, BuilderComponent],
   imports: [
     CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
     NetworkRoutingModule,
     FontAwesomeModule,
     MatPaginatorModule,
     MatTableModule,
-    ReactiveFormsModule,
     MatFormFieldModule,
     MatInputModule,
     MatCheckboxModule,